+++ ci/release/update-version.sh
@@ -47,9 +47,10 @@ function sed_runner() {
 }
 
 # RTD update
-sed_runner 's/release = ''.*''/release = '"'${NEXT_FULL_TAG}'"'/g' docs/source/conf.py
+sed_runner 's/version = .*/version = '"'${NEXT_SHORT_TAG}'"'/g' docs/source/conf.py
+sed_runner 's/release = .*/release = '"'${NEXT_FULL_TAG}'"'/g' docs/source/conf.py
 
 # bump rmm
 for FILE in conda/environments/*.yml; do
-  sed_runner "s/cudf>=.*/cudf>=${NEXT_SHORT_TAG}/g" ${FILE};
+  sed_runner "s/cudf=.*/cudf=${NEXT_SHORT_TAG}/g" ${FILE};
 done