+++ python/cuxfilter/tests/charts/core/test_core_widget.py
@@ -1,6 +1,7 @@
 import pytest
 
 from cuxfilter.charts.core.core_widget import BaseWidget
+from cuxfilter.layouts import chart_view
 
 
 class TestBaseWidget:
@@ -33,7 +34,7 @@ class TestBaseWidget:
         bw = BaseWidget("test_x")
         bw.chart = chart
 
-        assert bw.view() == _chart
+        assert str(bw.view()) == str(chart_view(_chart, width=bw.width))
 
     def test_add_event(self):
         bw = BaseWidget("test_x")